вынесла notification в отдельный файл
